feat: nil-based credit pricing

+++ nilcc-api/src/workload-tier/workload-tier.entity.ts
@@ -20,6 +20,6 @@ export class WorkloadTierEntity {
   @Column({ type: "int" })
   disk: number;
 
-  @Column({ type: "int" })
+  @Column({ type: "float" })
   cost: number;
 }
